fix: address Copilot review comments and regression fixes
- Host.Arch() falls back to rig live detection when k0sctl OSRelease
  field lacks arch (fixes OS_OVERRIDE and ID_LIKE smoke-test failures)
- Move Linux-specific host tests to host_posix_test.go with //go:build
  !windows to fix Windows CI failures from POSIX path assumptions
- daemon_reload: use h.K0sServiceName() instead of hardcoded
  "k0scontroller" so worker hosts reload the correct service
- apply_manifests: return error when kubectl apply fails instead of
  silently logging and continuing
- validate_hosts: fix fs.WalkDir called with glob pattern (globs are not
  expanded); walk the directory and filter by name prefix instead
- host: normalize backslashes in k0sBinaryPathDir() before path.Dir()
  to handle user-configured Windows-style paths correctly
